// dict/api/task/types.ts
/* eslint-disable */
// tslint:disable
/**
 * MICRO SERVICES CURD OPERATIONS
 * 
 *
 * OpenAPI spec version: 2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

import { Express } from 'express'
import { Api } from '../../models'

export interface TaskApi {
	getAnswer: (questionId: string | undefined) => Promise<GetAnswerResponse>
	createQuestions: (request: Api.QuestionData | undefined) => Promise<CreateQuestionsResponse>
	taskCreate: (request: Api.TaskData | undefined) => Promise<TaskCreateResponse>
	getGetAnswers: () => Promise<GetGetAnswersResponse>
	getGetQuestions: (topicid: string) => Promise<GetGetQuestionsResponse>
	getTask: () => Promise<GetTaskResponse>
	answerCreate: (request: Api.AnswerData | undefined) => Promise<AnswerCreateResponse>
}

export type GetAnswerResponse = GetAnswer200Response | GetAnswer400Response

export interface GetAnswer200Response {
	status: 200
	body: Api.AnswerData[]
	headers?: never
}

export interface GetAnswer400Response {
	status: 400
	body: Api.ApiResponse
	headers?: never
}

export type CreateQuestionsResponse = CreateQuestions201Response | CreateQuestions400Response

export interface CreateQuestions201Response {
	status: 201
	body: Api.ApiResponse
	headers?: never
}

export interface CreateQuestions400Response {
	status: 400
	body: Api.ApiResponse
	headers?: never
}

export type TaskCreateResponse = TaskCreate201Response | TaskCreate400Response

export interface TaskCreate201Response {
	status: 201
	body: Api.ApiResponse
	headers?: never
}

export interface TaskCreate400Response {
	status: 400
	body: Api.ApiResponse
	headers?: never
}

export type GetGetAnswersResponse = GetGetAnswers200Response | GetGetAnswers400Response

export interface GetGetAnswers200Response {
	status: 200
	body: Api.AnswerData[]
	headers?: never
}

export interface GetGetAnswers400Response {
	status: 400
	body: Api.ApiResponse
	headers?: never
}

export type GetGetQuestionsResponse = GetGetQuestions201Response | GetGetQuestions400Response

export interface GetGetQuestions201Response {
	status: 201
	body: Api.QuestionData[]
	headers?: never
}

export interface GetGetQuestions400Response {
	status: 400
	body: Api.ApiResponse
	headers?: never
}

export type GetTaskResponse = GetTask201Response | GetTask400Response

export interface GetTask201Response {
	status: 201
	body: Api.TaskData[]
	headers?: never
}

export interface GetTask400Response {
	status: 400
	body: Api.ApiResponse
	headers?: never
}

export type AnswerCreateResponse = AnswerCreate201Response | AnswerCreate400Response

export interface AnswerCreate201Response {
	status: 201
	body: Api.ApiResponse
	headers?: never
}

export interface AnswerCreate400Response {
	status: 400
	body: Api.ApiResponse
	headers?: never
}